rag/requirements2.txt
google-generativeai
python-dotenv
langchain-community
langchain-huggingface
chromadb
sentence-transformers
